manage_web/requirements.txt
Django

# For image field
Pillow
django-cleanup

django-nested-admin
django-cors-headers
djangorestframework
drf-writable-nested

# For drf filtering
django-filter

# For authentication
django-oauth-toolkit
social-auth-app-django
django-rest-framework-social-oauth2

# For API documentation
packaging    # drf yasg depends on packaging
coreapi      # drf yasg depends on coreapi
drf-yasg

# For mysql db
mysqlclient

# For using S3
boto3
django-storages

# For phonenumber
django-phonenumber-field
phonenumbers

# For deployment
gunicorn